fix: remove validateSelfSignedCert function
